@@ -1,10 +1,10 @@
 name: attributed_text
 description: Text with metadata spans for easy text editing and styling.
-version: 0.3.0
+version: 0.3.2
 homepage: https://github.com/superlistapp/super_editor
 
 environment:
-  sdk: ">=2.12.0 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   characters: ^1.2.0
