@@ -1,6 +1,5 @@
 import 'dart:math';
 
-import 'package:example_docs/editor.dart';
 import 'package:example_docs/infrastructure/icon_selector.dart';
 import 'package:example_docs/infrastructure/color_selector.dart';
 import 'package:example_docs/infrastructure/text_item_selector.dart';
@@ -42,7 +41,7 @@ class DocsEditorToolbar extends StatefulWidget {
 }
 
 class _DocsEditorToolbarState extends State<DocsEditorToolbar> {
-  /// Groups the aditional toolbar options popover, which is shown by tapping
+  /// Groups the additional toolbar options popover, which is shown by tapping
   /// the "more items" button with the popovers shown by the toolbar items,
   /// like the color picker.
   static const _tapRegionGroupId = 'docs_toolbar';
@@ -282,7 +281,7 @@ class _DocsEditorToolbarState extends State<DocsEditorToolbar> {
     ]);
 
     // Clear the field and hide the URL bar
-    _urlController!.clear();
+    _urlController!.clearTextAndSelection();
     _urlFocusNode.unfocus(disposition: UnfocusDisposition.previouslyFocusedChild);
     _linkPopoverController.close();
     setState(() {});
@@ -621,7 +620,7 @@ class _DocsEditorToolbarState extends State<DocsEditorToolbar> {
         height: 40,
         child: Padding(
           padding: const EdgeInsets.symmetric(vertical: 6.0, horizontal: 9.0),
-          child: _GroupedToolbarItens(
+          child: _GroupedToolbarItems(
             tapRegionGroupId: _tapRegionGroupId,
             visibleGroupCount: visibleGroupCount,
             groups: [
@@ -802,15 +801,15 @@ class _DocsEditorToolbarState extends State<DocsEditorToolbar> {
         ),
         label: const Text('Menus'),
         style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all(Colors.white),
-          overlayColor: MaterialStateProperty.all(Colors.transparent),
-          foregroundColor: MaterialStateProperty.all(Colors.black),
-          fixedSize: MaterialStateProperty.all(const Size(100, 30)),
-          minimumSize: MaterialStateProperty.all(const Size(100, 30)),
-          textStyle: MaterialStateProperty.all(
+          backgroundColor: WidgetStateProperty.all(Colors.white),
+          overlayColor: WidgetStateProperty.all(Colors.transparent),
+          foregroundColor: WidgetStateProperty.all(Colors.black),
+          fixedSize: WidgetStateProperty.all(const Size(100, 30)),
+          minimumSize: WidgetStateProperty.all(const Size(100, 30)),
+          textStyle: WidgetStateProperty.all(
             const TextStyle(fontWeight: FontWeight.w200),
           ),
-          mouseCursor: MaterialStateProperty.all(SystemMouseCursors.text),
+          mouseCursor: WidgetStateProperty.all(SystemMouseCursors.text),
         ),
       ),
     );
@@ -1188,7 +1187,7 @@ class _DocsEditorToolbarState extends State<DocsEditorToolbar> {
               onPressed: () {
                 setState(() {
                   _urlFocusNode.unfocus();
-                  _urlController!.clear();
+                  _urlController!.clearTextAndSelection();
                 });
               },
             ),
@@ -1231,8 +1230,8 @@ class _DocsEditorToolbarState extends State<DocsEditorToolbar> {
 /// Only the groups with index less than [visibleGroupCount]
 /// are displayed. When there is any hidden groups, a button is
 /// displayed to show a popover with the remaining groups.
-class _GroupedToolbarItens extends StatefulWidget {
-  const _GroupedToolbarItens({
+class _GroupedToolbarItems extends StatefulWidget {
+  const _GroupedToolbarItems({
     required this.groups,
     required this.visibleGroupCount,
     this.tapRegionGroupId,
@@ -1259,10 +1258,10 @@ class _GroupedToolbarItens extends StatefulWidget {
   final String? tapRegionGroupId;
 
   @override
-  State<_GroupedToolbarItens> createState() => _GroupedToolbarItensState();
+  State<_GroupedToolbarItems> createState() => _GroupedToolbarItemsState();
 }
 
-class _GroupedToolbarItensState extends State<_GroupedToolbarItens> {
+class _GroupedToolbarItemsState extends State<_GroupedToolbarItems> {
   final PopoverController _popoverController = PopoverController();
 
   @override
@@ -1435,19 +1434,19 @@ class ToolbarImageButton extends StatefulWidget {
 }
 
 class _ToolbarImageButtonState extends State<ToolbarImageButton> {
-  final MaterialStatesController _statesController = MaterialStatesController();
+  final WidgetStatesController _statesController = WidgetStatesController();
 
   @override
   void initState() {
     super.initState();
-    _statesController.update(MaterialState.selected, widget.selected);
+    _statesController.update(WidgetState.selected, widget.selected);
   }
 
   @override
   void didUpdateWidget(covariant ToolbarImageButton oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.selected != widget.selected) {
-      _statesController.update(MaterialState.selected, widget.selected);
+      _statesController.update(WidgetState.selected, widget.selected);
     }
   }
 
@@ -1467,9 +1466,9 @@ class _ToolbarImageButtonState extends State<ToolbarImageButton> {
         onPressed: widget.onPressed,
         statesController: _statesController,
         style: defaultToolbarButtonStyle.copyWith(
-          fixedSize: MaterialStateProperty.all(size),
-          minimumSize: MaterialStateProperty.all(size),
-          maximumSize: MaterialStateProperty.all(size),
+          fixedSize: WidgetStateProperty.all(size),
+          minimumSize: WidgetStateProperty.all(size),
+          maximumSize: WidgetStateProperty.all(size),
         ),
         child: widget.child,
       ),
