@@ -241,12 +241,12 @@ class _DrawerButton extends StatelessWidget {
       width: double.infinity,
       child: ElevatedButton(
         style: ButtonStyle(
-            backgroundColor: MaterialStateColor.resolveWith((states) {
+            backgroundColor: WidgetStateColor.resolveWith((states) {
               if (isSelected) {
                 return const Color(0xFFBBBBBB);
               }
 
-              if (states.contains(MaterialState.hovered)) {
+              if (states.contains(WidgetState.hovered)) {
                 return Colors.grey.withOpacity(0.1);
               }
 
@@ -254,9 +254,9 @@ class _DrawerButton extends StatelessWidget {
             }),
             // splashFactory: NoSplash.splashFactory,
             foregroundColor:
-                MaterialStateColor.resolveWith((states) => isSelected ? Colors.white : const Color(0xFFBBBBBB)),
-            elevation: MaterialStateProperty.resolveWith((states) => 0),
-            padding: MaterialStateProperty.resolveWith((states) => const EdgeInsets.all(16))),
+                WidgetStateColor.resolveWith((states) => isSelected ? Colors.white : const Color(0xFFBBBBBB)),
+            elevation: WidgetStateProperty.resolveWith((states) => 0),
+            padding: WidgetStateProperty.resolveWith((states) => const EdgeInsets.all(16))),
         onPressed: isSelected ? null : onPressed,
         child: Row(
           children: [
