@@ -99,22 +99,30 @@ class DocumentChangeLog {
 
 /// Marker interface for all document changes.
 abstract class DocumentChange {
-  // Marker interface
+  const DocumentChange();
+
+  /// Describes this change in a human-readable way.
+  String describe() => toString();
 }
 
 /// A [DocumentChange] that impacts a single, specified [DocumentNode] with [nodeId].
-abstract class NodeDocumentChange implements DocumentChange {
+abstract class NodeDocumentChange extends DocumentChange {
+  const NodeDocumentChange();
+
   String get nodeId;
 }
 
 /// A new [DocumentNode] was inserted in the [Document].
-class NodeInsertedEvent implements NodeDocumentChange {
+class NodeInsertedEvent extends NodeDocumentChange {
   const NodeInsertedEvent(this.nodeId, this.insertionIndex);
 
   @override
   final String nodeId;
 
   final int insertionIndex;
+
+  @override
+  String describe() => "Inserted node: $nodeId";
 
   @override
   String toString() => "NodeInsertedEvent ($nodeId)";
@@ -132,7 +140,7 @@ class NodeInsertedEvent implements NodeDocumentChange {
 }
 
 /// A [DocumentNode] was moved to a new index.
-class NodeMovedEvent implements NodeDocumentChange {
+class NodeMovedEvent extends NodeDocumentChange {
   const NodeMovedEvent({
     required this.nodeId,
     required this.from,
@@ -143,6 +151,9 @@ class NodeMovedEvent implements NodeDocumentChange {
   final String nodeId;
   final int from;
   final int to;
+
+  @override
+  String describe() => "Moved node ($nodeId): $from -> $to";
 
   @override
   String toString() => "NodeMovedEvent ($nodeId: $from -> $to)";
@@ -161,13 +172,16 @@ class NodeMovedEvent implements NodeDocumentChange {
 }
 
 /// A [DocumentNode] was removed from the [Document].
-class NodeRemovedEvent implements NodeDocumentChange {
+class NodeRemovedEvent extends NodeDocumentChange {
   const NodeRemovedEvent(this.nodeId, this.removedNode);
 
   @override
   final String nodeId;
 
   final DocumentNode removedNode;
+
+  @override
+  String describe() => "Removed node: $nodeId";
 
   @override
   String toString() => "NodeRemovedEvent ($nodeId)";
@@ -185,11 +199,14 @@ class NodeRemovedEvent implements NodeDocumentChange {
 /// A node change might signify a content change, such as text changing in a paragraph, or
 /// it might signify a node changing its type of content, such as converting a paragraph
 /// to an image.
-class NodeChangeEvent implements NodeDocumentChange {
+class NodeChangeEvent extends NodeDocumentChange {
   const NodeChangeEvent(this.nodeId);
 
   @override
   final String nodeId;
+
+  @override
+  String describe() => "Changed node: $nodeId";
 
   @override
   String toString() => "NodeChangeEvent ($nodeId)";
@@ -380,6 +397,8 @@ abstract class DocumentNode implements ChangeNotifier {
 
   /// Returns a copy of this node's metadata.
   Map<String, dynamic> copyMetadata() => Map.from(_metadata);
+
+  DocumentNode copy();
 
   @override
   bool operator ==(Object other) =>
