@@ -1,5 +1,7 @@
 import 'dart:ui';
 
+import 'package:super_editor/src/default_editor/text.dart';
+
 import 'document.dart';
 
 /// A selection within a [Document].
@@ -86,6 +88,20 @@ class DocumentSelection extends DocumentRange {
 
   @override
   String toString() {
+    if (base.nodeId == extent.nodeId) {
+      final basePosition = base.nodePosition;
+      final extentPosition = extent.nodePosition;
+      if (basePosition is TextNodePosition && extentPosition is TextNodePosition) {
+        if (basePosition.offset == extentPosition.offset) {
+          return "[Selection] - ${base.nodeId}: ${extentPosition.offset}";
+        }
+
+        return "[Selection] - ${base.nodeId}: [${basePosition.offset}, ${extentPosition.offset}]";
+      }
+
+      return "[Selection] - ${base.nodeId}: [${base.nodePosition}, ${extent.nodePosition}]";
+    }
+
     return '[DocumentSelection] - \n  base: ($base),\n  extent: ($extent)';
   }
 
@@ -283,7 +299,21 @@ class DocumentRange {
 
   @override
   String toString() {
-    return '[DocumentRange] - start: ($start), end: ($end)';
+    if (start.nodeId == end.nodeId) {
+      final startPosition = start.nodePosition;
+      final endPosition = end.nodePosition;
+      if (startPosition is TextNodePosition && endPosition is TextNodePosition) {
+        if (startPosition.offset == endPosition.offset) {
+          return "[Range] - ${start.nodeId}: ${endPosition.offset}";
+        }
+
+        return "[Range] - ${start.nodeId}: [${startPosition.offset}, ${endPosition.offset}]";
+      }
+
+      return "[Range] - ${start.nodeId}: [${start.nodePosition}, ${end.nodePosition}]";
+    }
+
+    return '[Range] - \n  start: ($start),\n  end: ($end)';
   }
 }
 
