@@ -238,12 +238,15 @@ class BlockquoteComponent extends StatelessWidget {
   }
 }
 
-class ConvertBlockquoteToParagraphCommand implements EditCommand {
+class ConvertBlockquoteToParagraphCommand extends EditCommand {
   ConvertBlockquoteToParagraphCommand({
     required this.nodeId,
   });
 
   final String nodeId;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -324,7 +327,7 @@ ExecutionInstruction splitBlockquoteWhenEnterPressed({
   return didSplit ? ExecutionInstruction.haltExecution : ExecutionInstruction.continueExecution;
 }
 
-class SplitBlockquoteCommand implements EditCommand {
+class SplitBlockquoteCommand extends EditCommand {
   SplitBlockquoteCommand({
     required this.nodeId,
     required this.splitPosition,
@@ -334,6 +337,9 @@ class SplitBlockquoteCommand implements EditCommand {
   final String nodeId;
   final TextPosition splitPosition;
   final String newNodeId;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
