@@ -14,6 +14,7 @@ import 'default_document_editor_reactions.dart';
 Editor createDefaultDocumentEditor({
   required MutableDocument document,
   required MutableDocumentComposer composer,
+  HistoryGroupingPolicy historyGroupingPolicy = defaultMergePolicy,
 }) {
   final editor = Editor(
     editables: {
@@ -21,6 +22,7 @@ Editor createDefaultDocumentEditor({
       Editor.composerKey: composer,
     },
     requestHandlers: List.from(defaultRequestHandlers),
+    historyGroupingPolicy: historyGroupingPolicy,
     reactionPipeline: List.from(defaultEditorReactions),
   );
 
@@ -51,6 +53,9 @@ final defaultRequestHandlers = List.unmodifiable(<EditRequestHandler>[
       : null,
   (request) => request is ChangeInteractionModeRequest //
       ? ChangeInteractionModeCommand(isInteractionModeDesired: request.isInteractionModeDesired)
+      : null,
+  (request) => request is RemoveComposerPreferenceStylesRequest //
+      ? RemoveComposerPreferenceStylesCommand(request.stylesToRemove)
       : null,
   (request) => request is InsertTextRequest
       ? InsertTextCommand(
@@ -134,6 +139,22 @@ final defaultRequestHandlers = List.unmodifiable(<EditRequestHandler>[
           alignment: request.alignment,
         )
       : null,
+  (request) => request is IndentParagraphRequest
+      ? IndentParagraphCommand(
+          request.nodeId,
+        )
+      : null,
+  (request) => request is UnIndentParagraphRequest
+      ? UnIndentParagraphCommand(
+          request.nodeId,
+        )
+      : null,
+  (request) => request is SetParagraphIndentRequest
+      ? SetParagraphIndentCommand(
+          request.nodeId,
+          level: request.level,
+        )
+      : null,
   (request) => request is ChangeParagraphBlockTypeRequest
       ? ChangeParagraphBlockTypeCommand(
           nodeId: request.nodeId,
@@ -155,6 +176,12 @@ final defaultRequestHandlers = List.unmodifiable(<EditRequestHandler>[
           isComplete: request.isComplete,
         )
       : null,
+  (request) => request is ConvertTaskToParagraphRequest
+      ? ConvertTaskToParagraphCommand(
+          nodeId: request.nodeId,
+          paragraphMetadata: request.paragraphMetadata,
+        )
+      : null,
   (request) => request is DeleteUpstreamAtBeginningOfNodeRequest && request.node is TaskNode
       ? ConvertTaskToParagraphCommand(nodeId: request.node.id, paragraphMetadata: request.node.metadata)
       : null,
@@ -163,6 +190,15 @@ final defaultRequestHandlers = List.unmodifiable(<EditRequestHandler>[
           nodeId: request.nodeId,
           isComplete: request.isComplete,
         )
+      : null,
+  (request) => request is IndentTaskRequest //
+      ? IndentTaskCommand(request.nodeId)
+      : null,
+  (request) => request is UnIndentTaskRequest //
+      ? UnIndentTaskCommand(request.nodeId)
+      : null,
+  (request) => request is SetTaskIndentRequest //
+      ? SetTaskIndentCommand(request.nodeId, request.indent)
       : null,
   (request) => request is SplitExistingTaskRequest
       ? SplitExistingTaskCommand(
@@ -216,7 +252,6 @@ final defaultRequestHandlers = List.unmodifiable(<EditRequestHandler>[
       ? PasteEditorCommand(
           content: request.content,
           pastePosition: request.pastePosition,
-          composer: request.composer,
         )
       : null,
 ]);
@@ -234,4 +269,6 @@ final defaultEditorReactions = List.unmodifiable([
   const ImageUrlConversionReaction(),
   const DashConversionReaction(),
   //---- End Content Conversions ---
+
+  UpdateSubTaskIndentAfterTaskDeletionReaction(),
 ]);
