@@ -1,3 +1,4 @@
+import 'dart:ui';
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
@@ -23,6 +24,7 @@ import 'package:super_editor/src/infrastructure/platforms/ios/floating_cursor.da
 import 'package:super_editor/src/infrastructure/platforms/ios/ios_document_controls.dart';
 import 'package:super_editor/src/infrastructure/platforms/ios/long_press_selection.dart';
 import 'package:super_editor/src/infrastructure/platforms/ios/magnifier.dart';
+import 'package:super_editor/src/infrastructure/platforms/ios/selection_heuristics.dart';
 import 'package:super_editor/src/infrastructure/platforms/mobile_documents.dart';
 import 'package:super_editor/src/infrastructure/platforms/platform.dart';
 import 'package:super_editor/src/infrastructure/signal_notifier.dart';
@@ -111,6 +113,7 @@ class SuperEditorIosControlsScope extends InheritedWidget {
 /// the caret, handles, floating cursor, magnifier, and toolbar.
 class SuperEditorIosControlsController {
   SuperEditorIosControlsController({
+    this.useIosSelectionHeuristics = true,
     this.handleColor,
     FloatingCursorController? floatingCursorController,
     this.magnifierBuilder,
@@ -124,6 +127,17 @@ class SuperEditorIosControlsController {
     _shouldShowMagnifier.dispose();
     _shouldShowToolbar.dispose();
   }
+
+  /// Whether to adjust the user's selection similar to the way iOS does.
+  ///
+  /// For example: iOS doesn't let users tap directly on a text offset. Instead,
+  /// iOS places the caret at the end of the word, or beginning of the word,
+  /// based on how close the user is to those locations when he taps.
+  ///
+  /// When this property is `true`, iOS-style heuristics should be used. When
+  /// this value is `false`, the user's gestures should directly impact the
+  /// area they touched.
+  final bool useIosSelectionHeuristics;
 
   /// Color of the text selection drag handles on iOS.
   final Color? handleColor;
@@ -304,6 +318,10 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
   bool get _isLongPressInProgress => _longPressStrategy != null;
   IosLongPressSelectionStrategy? _longPressStrategy;
 
+  // Cached view metrics to ignore unnecessary didChangeMetrics calls.
+  Size? _lastSize;
+  ViewPadding? _lastInsets;
+
   @override
   void initState() {
     super.initState();
@@ -330,6 +348,10 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
+
+    final view = View.of(context);
+    _lastSize = view.physicalSize;
+    _lastInsets = view.viewInsets;
 
     if (_controlsController != null) {
       _controlsController!.floatingCursorController.removeListener(_floatingCursorListener);
@@ -394,6 +416,21 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
 
   @override
   void didChangeMetrics() {
+    // DidChangeMetrics is sometimes called even when metrics doesn't change
+    // (i.e. on iOS with keyboard visible)
+    final view = View.of(context);
+    final size = view.physicalSize;
+    final insets = view.viewInsets;
+    if (size == _lastSize &&
+        _lastInsets?.left == insets.left &&
+        _lastInsets?.right == insets.right &&
+        _lastInsets?.top == insets.top &&
+        _lastInsets?.bottom == insets.bottom) {
+      return;
+    }
+    _lastSize = size;
+    _lastInsets = insets;
+
     // The available screen dimensions may have changed, e.g., due to keyboard
     // appearance/disappearance. Ensure the extent is still visible. Use a
     // post-frame callback to give the rest of the UI a chance to reflow, first.
@@ -492,11 +529,6 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
     return viewportBox.globalToLocal(globalOffset);
   }
 
-  Offset _viewportOffsetToDocumentOffset(Offset viewportOffset) {
-    final globalOffset = viewportBox.localToGlobal(viewportOffset);
-    return _docLayout.getDocumentOffsetFromAncestorOffset(globalOffset);
-  }
-
   RenderBox get interactorBox => context.findRenderObject() as RenderBox;
 
   /// Converts the given [interactorOffset] from the [DocumentInteractor]'s coordinate
@@ -591,6 +623,11 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
     widget.focusNode.requestFocus();
   }
 
+  void _onTapCancel() {
+    _tapDownLongPressTimer?.cancel();
+    _tapDownLongPressTimer = null;
+  }
+
   void _onTapUp(TapUpDetails details) {
     // Stop waiting for a long-press to start.
     _globalTapDownOffset = null;
@@ -641,9 +678,20 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
     }
 
     if (docPosition != null) {
+      late final DocumentPosition adjustedSelectionPosition;
+      if (docPosition.nodePosition is TextNodePosition) {
+        // The user tapped a text position. Adjust the position to the start
+        // or end of the word, as per iOS behavior.
+        adjustedSelectionPosition = _moveTapPositionToWordBoundary(docPosition);
+      } else {
+        // Selection isn't text. Don't adjust the position.
+        adjustedSelectionPosition = docPosition;
+      }
+
       final didTapOnExistingSelection = selection != null &&
           selection.isCollapsed &&
-          selection.extent.nodePosition.isEquivalentTo(docPosition.nodePosition);
+          selection.extent.nodeId == adjustedSelectionPosition.nodeId &&
+          selection.extent.nodePosition.isEquivalentTo(adjustedSelectionPosition.nodePosition);
 
       if (didTapOnExistingSelection) {
         // Toggle the toolbar display when the user taps on the collapsed caret,
@@ -654,7 +702,7 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
         _controlsController!.hideToolbar();
       }
 
-      final tappedComponent = _docLayout.getComponentByNodeId(docPosition.nodeId)!;
+      final tappedComponent = _docLayout.getComponentByNodeId(adjustedSelectionPosition.nodeId)!;
       if (!tappedComponent.isVisualSelectionSupported()) {
         // The user tapped a non-selectable component.
         // Place the document selection at the nearest selectable node
@@ -664,13 +712,21 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
           document: widget.document,
           documentLayoutResolver: widget.getDocumentLayout,
           currentSelection: widget.selection.value,
-          startingNode: widget.document.getNodeById(docPosition.nodeId)!,
+          startingNode: widget.document.getNodeById(adjustedSelectionPosition.nodeId)!,
         );
         return;
       } else {
         // Place the document selection at the location where the
         // user tapped.
-        _selectPosition(docPosition);
+        _selectPosition(adjustedSelectionPosition);
+      }
+
+      if (didTapOnExistingSelection) {
+        // The user tapped on the existing selection. Show the software keyboard.
+        //
+        // If the user didn't tap on an existing selection, the software keyboard will
+        // already be visible.
+        widget.openSoftwareKeyboard();
       }
 
       if (didTapOnExistingSelection) {
@@ -688,6 +744,25 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
     }
 
     widget.focusNode.requestFocus();
+  }
+
+  DocumentPosition _moveTapPositionToWordBoundary(DocumentPosition docPosition) {
+    if (!SuperEditorIosControlsScope.rootOf(context).useIosSelectionHeuristics) {
+      // iOS-style adjustments aren't desired. Don't adjust th given position.
+      return docPosition;
+    }
+
+    final text = (widget.document.getNodeById(docPosition.nodeId) as TextNode).text.text;
+    final tapOffset = (docPosition.nodePosition as TextNodePosition).offset;
+    if (tapOffset == text.length) {
+      return docPosition;
+    }
+    final adjustedSelectionOffset = IosHeuristics.adjustTapOffset(text, tapOffset);
+
+    return DocumentPosition(
+      nodeId: docPosition.nodeId,
+      nodePosition: TextNodePosition(offset: adjustedSelectionOffset),
+    );
   }
 
   void _onDoubleTapUp(TapUpDetails details) {
@@ -1301,6 +1376,7 @@ class _IosDocumentTouchInteractorState extends State<IosDocumentTouchInteractor>
           (TapSequenceGestureRecognizer recognizer) {
             recognizer
               ..onTapDown = _onTapDown
+              ..onTapCancel = _onTapCancel
               ..onTapUp = _onTapUp
               ..onDoubleTapUp = _onDoubleTapUp
               ..onTripleTapUp = _onTripleTapUp
@@ -1839,9 +1915,16 @@ class SuperEditorIosToolbarFocalPointDocumentLayerBuilder implements SuperEditor
 class SuperEditorIosHandlesDocumentLayerBuilder implements SuperEditorLayerBuilder {
   const SuperEditorIosHandlesDocumentLayerBuilder({
     this.handleColor,
+    this.caretWidth,
+    this.handleBallDiameter,
   });
 
   final Color? handleColor;
+  final double? caretWidth;
+
+  /// The diameter of the small circle that appears on the top and bottom of
+  /// expanded iOS text handles.
+  final double? handleBallDiameter;
 
   @override
   ContentLayerWidget build(BuildContext context, SuperEditorContext editContext) {
@@ -1864,6 +1947,8 @@ class SuperEditorIosHandlesDocumentLayerBuilder implements SuperEditorLayerBuild
       handleColor: handleColor ??
           SuperEditorIosControlsScope.maybeRootOf(context)?.handleColor ??
           Theme.of(context).primaryColor,
+      caretWidth: caretWidth ?? 2,
+      handleBallDiameter: handleBallDiameter ?? defaultIosHandleBallDiameter,
       shouldCaretBlink: SuperEditorIosControlsScope.rootOf(context).shouldCaretBlink,
       floatingCursorController: SuperEditorIosControlsScope.rootOf(context).floatingCursorController,
     );
@@ -1874,3 +1959,7 @@ const defaultIosMagnifierEnterAnimationDuration = Duration(milliseconds: 180);
 const defaultIosMagnifierExitAnimationDuration = Duration(milliseconds: 150);
 const defaultIosMagnifierAnimationCurve = Curves.easeInOut;
 const defaultIosMagnifierSize = Size(133, 96);
+
+/// The diameter of the small circle that appears on the top and bottom of
+/// expanded iOS text handles in dip.
+const defaultIosHandleBallDiameter = 16.0;
