@@ -4,7 +4,9 @@ import 'package:flutter/services.dart';
 import 'package:super_editor/src/core/document_composer.dart';
 import 'package:super_editor/src/core/edit_context.dart';
 import 'package:super_editor/src/core/editor.dart';
+import 'package:super_editor/src/core/styles.dart';
 import 'package:super_editor/src/default_editor/attributions.dart';
+import 'package:super_editor/src/default_editor/blocks/indentation.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
 import 'package:super_editor/src/infrastructure/attributed_text_styles.dart';
 import 'package:super_editor/src/infrastructure/keyboard.dart';
@@ -82,6 +84,17 @@ class ListItemNode extends TextNode {
   }
 
   @override
+  ListItemNode copy() {
+    return ListItemNode(
+      id: id,
+      text: text.copyText(0),
+      itemType: type,
+      indent: indent,
+      metadata: Map.from(metadata),
+    );
+  }
+
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       super == other &&
@@ -112,61 +125,57 @@ class ListItemComponentBuilder implements ComponentBuilder {
 
     int? ordinalValue;
     if (node.type == ListItemType.ordered) {
-      // Counts the number of ordered list items above the current node with the same indentation level. Ordered
-      // list items with the same indentation level might be separated by ordered or unordered list items with
-      // different indentation levels.
-      ordinalValue = 1;
-      DocumentNode? nodeAbove = document.getNodeBefore(node);
-      while (nodeAbove != null && nodeAbove is ListItemNode && nodeAbove.indent >= node.indent) {
-        if (nodeAbove.indent == node.indent) {
-          if (nodeAbove.type != ListItemType.ordered) {
-            // We found an unordered list item with the same indentation level as the ordered list item.
-            // Other ordered list items aboce this one do not belong to the same list.
-            break;
-          }
-          ordinalValue = ordinalValue! + 1;
-        }
-        nodeAbove = document.getNodeBefore(nodeAbove);
-      }
+      ordinalValue = computeListItemOrdinalValue(node, document);
     }
 
-    return ListItemComponentViewModel(
-      nodeId: node.id,
-      type: node.type,
-      indent: node.indent,
-      ordinalValue: ordinalValue,
-      text: node.text,
-      textStyleBuilder: noStyleBuilder,
-      selectionColor: const Color(0x00000000),
-    );
+    return switch (node.type) {
+      ListItemType.unordered => UnorderedListItemComponentViewModel(
+          nodeId: node.id,
+          indent: node.indent,
+          text: node.text,
+          textStyleBuilder: noStyleBuilder,
+          selectionColor: const Color(0x00000000),
+        ),
+      ListItemType.ordered => OrderedListItemComponentViewModel(
+          nodeId: node.id,
+          indent: node.indent,
+          ordinalValue: ordinalValue,
+          text: node.text,
+          textStyleBuilder: noStyleBuilder,
+          selectionColor: const Color(0x00000000),
+        ),
+    };
   }
 
   @override
   Widget? createComponent(
       SingleColumnDocumentComponentContext componentContext, SingleColumnLayoutComponentViewModel componentViewModel) {
-    if (componentViewModel is! ListItemComponentViewModel) {
+    if (componentViewModel is! UnorderedListItemComponentViewModel &&
+        componentViewModel is! OrderedListItemComponentViewModel) {
       return null;
     }
 
-    if (componentViewModel.type == ListItemType.unordered) {
+    if (componentViewModel is UnorderedListItemComponentViewModel) {
       return UnorderedListItemComponent(
         componentKey: componentContext.componentKey,
         text: componentViewModel.text,
         styleBuilder: componentViewModel.textStyleBuilder,
         indent: componentViewModel.indent,
+        dotStyle: componentViewModel.dotStyle,
         textSelection: componentViewModel.selection,
         selectionColor: componentViewModel.selectionColor,
         highlightWhenEmpty: componentViewModel.highlightWhenEmpty,
         composingRegion: componentViewModel.composingRegion,
         showComposingUnderline: componentViewModel.showComposingUnderline,
       );
-    } else if (componentViewModel.type == ListItemType.ordered) {
+    } else if (componentViewModel is OrderedListItemComponentViewModel) {
       return OrderedListItemComponent(
         componentKey: componentContext.componentKey,
         indent: componentViewModel.indent,
         listIndex: componentViewModel.ordinalValue!,
         text: componentViewModel.text,
         styleBuilder: componentViewModel.textStyleBuilder,
+        numeralStyle: componentViewModel.numeralStyle,
         textSelection: componentViewModel.selection,
         selectionColor: componentViewModel.selectionColor,
         highlightWhenEmpty: componentViewModel.highlightWhenEmpty,
@@ -181,13 +190,11 @@ class ListItemComponentBuilder implements ComponentBuilder {
   }
 }
 
-class ListItemComponentViewModel extends SingleColumnLayoutComponentViewModel with TextComponentViewModel {
+abstract class ListItemComponentViewModel extends SingleColumnLayoutComponentViewModel with TextComponentViewModel {
   ListItemComponentViewModel({
     required String nodeId,
     double? maxWidth,
     EdgeInsetsGeometry padding = EdgeInsets.zero,
-    required this.type,
-    this.ordinalValue,
     required this.indent,
     required this.text,
     required this.textStyleBuilder,
@@ -200,8 +207,6 @@ class ListItemComponentViewModel extends SingleColumnLayoutComponentViewModel wi
     this.showComposingUnderline = false,
   }) : super(nodeId: nodeId, maxWidth: maxWidth, padding: padding);
 
-  ListItemType type;
-  int? ordinalValue;
   int indent;
 
   @override
@@ -224,14 +229,129 @@ class ListItemComponentViewModel extends SingleColumnLayoutComponentViewModel wi
   bool showComposingUnderline;
 
   @override
-  ListItemComponentViewModel copy() {
-    return ListItemComponentViewModel(
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      super == other &&
+          other is ListItemComponentViewModel &&
+          runtimeType == other.runtimeType &&
+          nodeId == other.nodeId &&
+          indent == other.indent &&
+          text == other.text &&
+          textDirection == other.textDirection &&
+          selection == other.selection &&
+          selectionColor == other.selectionColor &&
+          composingRegion == other.composingRegion &&
+          showComposingUnderline == other.showComposingUnderline;
+
+  @override
+  int get hashCode =>
+      super.hashCode ^
+      nodeId.hashCode ^
+      indent.hashCode ^
+      text.hashCode ^
+      textDirection.hashCode ^
+      selection.hashCode ^
+      selectionColor.hashCode ^
+      composingRegion.hashCode;
+}
+
+class UnorderedListItemComponentViewModel extends ListItemComponentViewModel {
+  UnorderedListItemComponentViewModel({
+    required super.nodeId,
+    super.maxWidth,
+    super.padding = EdgeInsets.zero,
+    required super.indent,
+    required super.text,
+    required super.textStyleBuilder,
+    this.dotStyle = const ListItemDotStyle(),
+    super.textDirection = TextDirection.ltr,
+    super.textAlignment = TextAlign.left,
+    super.selection,
+    required super.selectionColor,
+    super.highlightWhenEmpty = false,
+    super.composingRegion,
+    super.showComposingUnderline = false,
+  });
+
+  ListItemDotStyle dotStyle = const ListItemDotStyle();
+
+  @override
+  void applyStyles(Map<String, dynamic> styles) {
+    super.applyStyles(styles);
+    dotStyle = ListItemDotStyle(
+      color: styles[Styles.dotColor],
+      shape: styles[Styles.dotShape] ?? BoxShape.circle,
+      size: styles[Styles.dotSize] ?? const Size(4, 4),
+    );
+  }
+
+  @override
+  UnorderedListItemComponentViewModel copy() {
+    return UnorderedListItemComponentViewModel(
       nodeId: nodeId,
       maxWidth: maxWidth,
       padding: padding,
-      type: type,
-      ordinalValue: ordinalValue,
       indent: indent,
+      text: text,
+      textStyleBuilder: textStyleBuilder,
+      dotStyle: dotStyle,
+      textDirection: textDirection,
+      selection: selection,
+      selectionColor: selectionColor,
+      composingRegion: composingRegion,
+      showComposingUnderline: showComposingUnderline,
+    );
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      super == other &&
+          other is UnorderedListItemComponentViewModel &&
+          runtimeType == other.runtimeType &&
+          dotStyle == other.dotStyle;
+
+  @override
+  int get hashCode => super.hashCode ^ dotStyle.hashCode;
+}
+
+class OrderedListItemComponentViewModel extends ListItemComponentViewModel {
+  OrderedListItemComponentViewModel({
+    required super.nodeId,
+    super.maxWidth,
+    super.padding = EdgeInsets.zero,
+    required super.indent,
+    this.ordinalValue,
+    this.numeralStyle = OrderedListNumeralStyle.arabic,
+    required super.text,
+    required super.textStyleBuilder,
+    super.textDirection = TextDirection.ltr,
+    super.textAlignment = TextAlign.left,
+    super.selection,
+    required super.selectionColor,
+    super.highlightWhenEmpty = false,
+    super.composingRegion,
+    super.showComposingUnderline = false,
+  });
+
+  final int? ordinalValue;
+  OrderedListNumeralStyle numeralStyle;
+
+  @override
+  void applyStyles(Map<String, dynamic> styles) {
+    super.applyStyles(styles);
+    numeralStyle = styles[Styles.listNumeralStyle] ?? OrderedListNumeralStyle.arabic;
+  }
+
+  @override
+  OrderedListItemComponentViewModel copy() {
+    return OrderedListItemComponentViewModel(
+      nodeId: nodeId,
+      maxWidth: maxWidth,
+      padding: padding,
+      indent: indent,
+      ordinalValue: ordinalValue,
+      numeralStyle: numeralStyle,
       text: text,
       textStyleBuilder: textStyleBuilder,
       textDirection: textDirection,
@@ -246,32 +366,37 @@ class ListItemComponentViewModel extends SingleColumnLayoutComponentViewModel wi
   bool operator ==(Object other) =>
       identical(this, other) ||
       super == other &&
-          other is ListItemComponentViewModel &&
+          other is OrderedListItemComponentViewModel &&
           runtimeType == other.runtimeType &&
-          nodeId == other.nodeId &&
-          type == other.type &&
           ordinalValue == other.ordinalValue &&
-          indent == other.indent &&
-          text == other.text &&
-          textDirection == other.textDirection &&
-          selection == other.selection &&
-          selectionColor == other.selectionColor &&
-          composingRegion == other.composingRegion &&
-          showComposingUnderline == other.showComposingUnderline;
+          numeralStyle == other.numeralStyle;
 
   @override
-  int get hashCode =>
-      super.hashCode ^
-      nodeId.hashCode ^
-      type.hashCode ^
-      ordinalValue.hashCode ^
-      indent.hashCode ^
-      text.hashCode ^
-      textDirection.hashCode ^
-      selection.hashCode ^
-      selectionColor.hashCode ^
-      composingRegion.hashCode ^
-      showComposingUnderline.hashCode;
+  int get hashCode => super.hashCode ^ ordinalValue.hashCode ^ numeralStyle.hashCode;
+}
+
+class ListItemDotStyle {
+  const ListItemDotStyle({
+    this.color,
+    this.shape = BoxShape.circle,
+    this.size = const Size(4, 4),
+  });
+
+  final Color? color;
+  final BoxShape shape;
+  final Size size;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is ListItemDotStyle &&
+          runtimeType == other.runtimeType &&
+          color == other.color &&
+          shape == other.shape &&
+          size == other.size;
+
+  @override
+  int get hashCode => super.hashCode ^ color.hashCode ^ shape.hashCode ^ size.hashCode;
 }
 
 /// Displays a un-ordered list item in a document.
@@ -284,8 +409,9 @@ class UnorderedListItemComponent extends StatefulWidget {
     required this.text,
     required this.styleBuilder,
     this.dotBuilder = _defaultUnorderedListItemDotBuilder,
+    this.dotStyle,
     this.indent = 0,
-    this.indentCalculator = _defaultIndentCalculator,
+    this.indentCalculator = defaultListItemIndentCalculator,
     this.textSelection,
     this.selectionColor = Colors.lightBlueAccent,
     this.showCaret = false,
@@ -300,6 +426,7 @@ class UnorderedListItemComponent extends StatefulWidget {
   final AttributedText text;
   final AttributionStyleBuilder styleBuilder;
   final UnorderedListItemDotBuilder dotBuilder;
+  final ListItemDotStyle? dotStyle;
   final int indent;
   final double Function(TextStyle, int indent) indentCalculator;
   final TextSelection? textSelection;
@@ -376,6 +503,24 @@ class _UnorderedListItemComponentState extends State<UnorderedListItemComponent>
   }
 }
 
+/// The styling of an ordered list numberal.
+enum OrderedListNumeralStyle {
+  /// Arabic numeral style (e.g. 1, 2, 3, ...).
+  arabic,
+
+  /// Lowercase alphabetic numeral style (e.g. a, b, c, ...).
+  lowerAlpha,
+
+  /// Uppercase alphabetic numeral style (e.g. A, B, C, ...).
+  upperAlpha,
+
+  /// Lowercase Roman numeral style (e.g. i, ii, iii, ...).
+  lowerRoman,
+
+  /// Uppercase Roman numeral style (e.g. I, II, III, ...).
+  upperRoman,
+}
+
 typedef UnorderedListItemDotBuilder = Widget Function(BuildContext, UnorderedListItemComponent);
 
 Widget _defaultUnorderedListItemDotBuilder(BuildContext context, UnorderedListItemComponent component) {
@@ -384,6 +529,8 @@ Widget _defaultUnorderedListItemDotBuilder(BuildContext context, UnorderedListIt
   // of the first character to determine the text style.
   final attributions = component.text.getAllAttributionsAt(0).toSet();
   final textStyle = component.styleBuilder(attributions);
+
+  final dotSize = component.dotStyle?.size ?? const Size(4, 4);
 
   return Align(
     alignment: Alignment.centerRight,
@@ -398,12 +545,12 @@ Widget _defaultUnorderedListItemDotBuilder(BuildContext context, UnorderedListIt
           WidgetSpan(
             alignment: PlaceholderAlignment.middle,
             child: Container(
-              width: 4,
-              height: 4,
+              width: dotSize.width,
+              height: dotSize.height,
               margin: const EdgeInsets.only(right: 10),
               decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                color: textStyle.color,
+                shape: component.dotStyle?.shape ?? BoxShape.circle,
+                color: component.dotStyle?.color ?? textStyle.color,
               ),
             ),
           ),
@@ -426,8 +573,9 @@ class OrderedListItemComponent extends StatefulWidget {
     required this.text,
     required this.styleBuilder,
     this.numeralBuilder = _defaultOrderedListItemNumeralBuilder,
+    this.numeralStyle = OrderedListNumeralStyle.arabic,
     this.indent = 0,
-    this.indentCalculator = _defaultIndentCalculator,
+    this.indentCalculator = defaultListItemIndentCalculator,
     this.textSelection,
     this.selectionColor = Colors.lightBlueAccent,
     this.showCaret = false,
@@ -443,8 +591,9 @@ class OrderedListItemComponent extends StatefulWidget {
   final AttributedText text;
   final AttributionStyleBuilder styleBuilder;
   final OrderedListItemNumeralBuilder numeralBuilder;
+  final OrderedListNumeralStyle numeralStyle;
   final int indent;
-  final double Function(TextStyle, int indent) indentCalculator;
+  final TextBlockIndentCalculator indentCalculator;
   final TextSelection? textSelection;
   final Color selectionColor;
   final bool showCaret;
@@ -522,7 +671,8 @@ class _OrderedListItemComponentState extends State<OrderedListItemComponent> {
 
 typedef OrderedListItemNumeralBuilder = Widget Function(BuildContext, OrderedListItemComponent);
 
-double _defaultIndentCalculator(TextStyle textStyle, int indent) {
+/// The standard [TextBlockIndentCalculator] used by list items in `SuperEditor`.
+double defaultListItemIndentCalculator(TextStyle textStyle, int indent) {
   return (textStyle.fontSize! * 0.60) * 4 * (indent + 1);
 }
 
@@ -541,13 +691,130 @@ Widget _defaultOrderedListItemNumeralBuilder(BuildContext context, OrderedListIt
       child: Padding(
         padding: const EdgeInsets.only(right: 5.0),
         child: Text(
-          '${component.listIndex}.',
+          '${_numeralForIndex(component.listIndex, component.numeralStyle)}.',
           textAlign: TextAlign.right,
           style: textStyle,
         ),
       ),
     ),
   );
+}
+
+/// Returns the text to be displayed for the given [numeral] and [numeralStyle].
+String _numeralForIndex(int numeral, OrderedListNumeralStyle numeralStyle) {
+  return switch (numeralStyle) {
+    OrderedListNumeralStyle.arabic => '$numeral',
+    OrderedListNumeralStyle.upperRoman => _intToRoman(numeral) ?? '$numeral',
+    OrderedListNumeralStyle.lowerRoman => _intToRoman(numeral)?.toLowerCase() ?? '$numeral',
+    OrderedListNumeralStyle.upperAlpha => _intToAlpha(numeral),
+    OrderedListNumeralStyle.lowerAlpha => _intToAlpha(numeral).toLowerCase(),
+  };
+}
+
+/// Converts a number to its Roman numeral representation.
+///
+/// Returns `null` if the number is greater than 3999, as we don't support the
+/// vinculum notation. See more at https://en.wikipedia.org/wiki/Roman_numerals#cite_ref-Ifrah2000_52-1.
+String? _intToRoman(int number) {
+  if (number <= 0) {
+    throw ArgumentError('Roman numerals are only defined for positive integers');
+  }
+
+  if (number > 3999) {
+    // Starting from 4000, the Roman numeral representation uses a bar over the numeral to represent
+    // a multiplication by 1000. We don't support this notation.
+    return null;
+  }
+
+  const values = [1000, 500, 100, 50, 10, 5, 1];
+  const symbols = ["M", "D", "C", "L", "X", "V", "I"];
+
+  int remainingValueToConvert = number;
+
+  final result = StringBuffer();
+
+  for (int i = 0; i < values.length; i++) {
+    final currentSymbol = symbols[i];
+    final currentSymbolValue = values[i];
+
+    final count = remainingValueToConvert ~/ currentSymbolValue;
+
+    if (count > 0 && count < 4) {
+      // The number is bigger than the current symbol's value. Add the appropriate
+      // number of digits, respecting the maximum of three consecutive symbols.
+      // For example, for 300 we would add "CCC", but for 400 we won't add "CCCC".
+      result.write(currentSymbol * count);
+
+      remainingValueToConvert %= currentSymbolValue;
+    }
+
+    if (remainingValueToConvert <= 0) {
+      // The conversion is complete.
+      break;
+    }
+
+    // We still have some value to convert. Check if we can use subtractive notation.
+    if (i % 2 == 0 && i + 2 < values.length) {
+      // Numbers in even positions (0, 2, 4) can be subtracted with other numbers
+      // two positions to the right of them:
+      //
+      //  - 1000 (M) can be subtracted with 100 (C).
+      //  - 100 (C) can be subtracted with 10 (X).
+      //  - 10 (X) can be subtracted with 1 (I).
+      //
+      // Check if we can do this subtraction.
+      final subtractiveValue = currentSymbolValue - values[i + 2];
+      if (remainingValueToConvert >= subtractiveValue) {
+        result.write(symbols[i + 2] + currentSymbol);
+        remainingValueToConvert -= subtractiveValue;
+      }
+    } else if (i % 2 != 0 && i + 1 < values.length) {
+      // Numbers in odd positions (1, 3, 5) can be subtracted with the number
+      // immediately after it to the right:
+      //
+      // - 500 (D) can be subtracted with 100 (C).
+      // - 50 (L) can be subtracted with 10 (X).
+      // - 5 (V) can be subtracted with 1 (I).
+      //
+      // Check if we can do this subtraction.
+      final subtractiveValue = currentSymbolValue - values[i + 1];
+      if (remainingValueToConvert >= subtractiveValue) {
+        result.write(symbols[i + 1] + currentSymbol);
+        remainingValueToConvert -= subtractiveValue;
+      }
+    }
+  }
+
+  return result.toString();
+}
+
+/// Converts a number to a string composed of A-Z characters.
+///
+/// For example:
+/// - 1 -> A
+/// - 2 -> B
+/// - ...
+/// - 26 -> Z
+/// - 27 -> AA
+/// - 28 -> AB
+String _intToAlpha(int num) {
+  const characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+  const base = characters.length;
+
+  String result = '';
+
+  while (num > 0) {
+    // Convert to 0-based index.
+    num -= 1;
+
+    // Find the next character to be added.
+    result = characters[num % base] + result;
+
+    // Move to the next digit.
+    num = num ~/ base;
+  }
+
+  return result;
 }
 
 class IndentListItemRequest implements EditRequest {
@@ -558,12 +825,15 @@ class IndentListItemRequest implements EditRequest {
   final String nodeId;
 }
 
-class IndentListItemCommand implements EditCommand {
+class IndentListItemCommand extends EditCommand {
   IndentListItemCommand({
     required this.nodeId,
   });
 
   final String nodeId;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -593,12 +863,15 @@ class UnIndentListItemRequest implements EditRequest {
   final String nodeId;
 }
 
-class UnIndentListItemCommand implements EditCommand {
+class UnIndentListItemCommand extends EditCommand {
   UnIndentListItemCommand({
     required this.nodeId,
   });
 
   final String nodeId;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -635,7 +908,7 @@ class ConvertListItemToParagraphRequest implements EditRequest {
   final Map<String, dynamic>? paragraphMetadata;
 }
 
-class ConvertListItemToParagraphCommand implements EditCommand {
+class ConvertListItemToParagraphCommand extends EditCommand {
   ConvertListItemToParagraphCommand({
     required this.nodeId,
     this.paragraphMetadata,
@@ -643,6 +916,9 @@ class ConvertListItemToParagraphCommand implements EditCommand {
 
   final String nodeId;
   final Map<String, dynamic>? paragraphMetadata;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -679,7 +955,7 @@ class ConvertParagraphToListItemRequest implements EditRequest {
   final ListItemType type;
 }
 
-class ConvertParagraphToListItemCommand implements EditCommand {
+class ConvertParagraphToListItemCommand extends EditCommand {
   ConvertParagraphToListItemCommand({
     required this.nodeId,
     required this.type,
@@ -687,6 +963,9 @@ class ConvertParagraphToListItemCommand implements EditCommand {
 
   final String nodeId;
   final ListItemType type;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -719,7 +998,7 @@ class ChangeListItemTypeRequest implements EditRequest {
   final ListItemType newType;
 }
 
-class ChangeListItemTypeCommand implements EditCommand {
+class ChangeListItemTypeCommand extends EditCommand {
   ChangeListItemTypeCommand({
     required this.nodeId,
     required this.newType,
@@ -727,6 +1006,9 @@ class ChangeListItemTypeCommand implements EditCommand {
 
   final String nodeId;
   final ListItemType newType;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -760,7 +1042,7 @@ class SplitListItemRequest implements EditRequest {
   final String newNodeId;
 }
 
-class SplitListItemCommand implements EditCommand {
+class SplitListItemCommand extends EditCommand {
   SplitListItemCommand({
     required this.nodeId,
     required this.splitPosition,
@@ -770,6 +1052,9 @@ class SplitListItemCommand implements EditCommand {
   final String nodeId;
   final TextPosition splitPosition;
   final String newNodeId;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -928,4 +1213,32 @@ ExecutionInstruction splitListItemWhenEnterPressed({
 
   final didSplitListItem = editContext.commonOps.insertBlockLevelNewline();
   return didSplitListItem ? ExecutionInstruction.haltExecution : ExecutionInstruction.continueExecution;
+}
+
+/// Computes the ordinal value of an ordered list item.
+///
+/// Walks backwards counting the number of ordered list items above the [listItem] with the same indentation level.
+///
+/// The ordinal value starts at 1.
+int computeListItemOrdinalValue(ListItemNode listItem, Document document) {
+  if (listItem.type != ListItemType.ordered) {
+    // Unordered list items do not have an ordinal value.
+    return 0;
+  }
+
+  int ordinalValue = 1;
+  DocumentNode? nodeAbove = document.getNodeBefore(listItem);
+  while (nodeAbove != null && nodeAbove is ListItemNode && nodeAbove.indent >= listItem.indent) {
+    if (nodeAbove.indent == listItem.indent) {
+      if (nodeAbove.type != ListItemType.ordered) {
+        // We found an unordered list item with the same indentation level as the ordered list item.
+        // Other ordered list items above this one do not belong to the same list.
+        break;
+      }
+      ordinalValue = ordinalValue + 1;
+    }
+    nodeAbove = document.getNodeBefore(nodeAbove);
+  }
+
+  return ordinalValue;
 }
