@@ -28,7 +28,7 @@ class PasteStructuredContentEditorRequest implements EditRequest {
 
 /// Inserts given structured content, in the form of a `List` of [DocumentNode]s at a
 /// given paste position within the document.
-class PasteStructuredContentEditorCommand implements EditCommand {
+class PasteStructuredContentEditorCommand extends EditCommand {
   PasteStructuredContentEditorCommand({
     required List<DocumentNode> content,
     required DocumentPosition pastePosition,
@@ -37,6 +37,9 @@ class PasteStructuredContentEditorCommand implements EditCommand {
 
   final List<DocumentNode> _content;
   final DocumentPosition _pastePosition;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -73,6 +76,19 @@ class PasteStructuredContentEditorCommand implements EditCommand {
           // Only text nodes are merge-able, therefore we know that the first pasted node
           // is a TextNode.
           textToInsert: (pastedNode as TextNode).text,
+        ),
+      );
+      executor.executeCommand(
+        ChangeSelectionCommand(
+          DocumentSelection.collapsed(
+            position: DocumentPosition(
+              nodeId: pastePosition.nodeId,
+              nodePosition: TextNodePosition(
+                  offset: (pastePosition.nodePosition as TextNodePosition).offset + pastedNode.text.length),
+            ),
+          ),
+          SelectionChangeType.insertContent,
+          SelectionReason.userInteraction,
         ),
       );
 
@@ -270,6 +286,9 @@ class InsertNodeAtIndexCommand extends EditCommand {
 
   final int nodeIndex;
   final DocumentNode newNode;
+
+  @override
+  String describe() => "Insert node at index $nodeIndex: $newNode";
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -600,12 +619,15 @@ class ReplaceNodeWithEmptyParagraphWithCaretRequest implements EditRequest {
   int get hashCode => nodeId.hashCode;
 }
 
-class ReplaceNodeWithEmptyParagraphWithCaretCommand implements EditCommand {
+class ReplaceNodeWithEmptyParagraphWithCaretCommand extends EditCommand {
   ReplaceNodeWithEmptyParagraphWithCaretCommand({
     required this.nodeId,
   });
 
   final String nodeId;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -653,12 +675,18 @@ class DeleteContentRequest implements EditRequest {
   final DocumentRange documentRange;
 }
 
-class DeleteContentCommand implements EditCommand {
+class DeleteContentCommand extends EditCommand {
   DeleteContentCommand({
     required this.documentRange,
   });
 
   final DocumentRange documentRange;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
+
+  @override
+  String describe() => "Delete content within range: $documentRange";
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
@@ -1030,12 +1058,15 @@ class DeleteNodeRequest implements EditRequest {
   final String nodeId;
 }
 
-class DeleteNodeCommand implements EditCommand {
+class DeleteNodeCommand extends EditCommand {
   DeleteNodeCommand({
     required this.nodeId,
   });
 
   final String nodeId;
+
+  @override
+  HistoryBehavior get historyBehavior => HistoryBehavior.undoable;
 
   @override
   void execute(EditContext context, CommandExecutor executor) {
