@@ -33,6 +33,7 @@ import 'package:super_editor/src/infrastructure/platforms/mac/mac_ime.dart';
 import 'package:super_editor/src/infrastructure/platforms/platform.dart';
 import 'package:super_editor/src/infrastructure/signal_notifier.dart';
 import 'package:super_editor/src/infrastructure/text_input.dart';
+import 'package:super_editor/src/undo_redo.dart';
 import 'package:super_text_layout/super_text_layout.dart';
 
 import '../infrastructure/document_gestures_interaction_overrides.dart';
@@ -138,7 +139,7 @@ class SuperEditor extends StatefulWidget {
         selectionStyles = selectionStyle ?? defaultSelectionStyle,
         componentBuilders = componentBuilders != null
             ? [...componentBuilders, const UnknownComponentBuilder()]
-            : [...defaultComponentBuilders, const UnknownComponentBuilder()],
+            : [...defaultComponentBuilders, TaskComponentBuilder(editor), const UnknownComponentBuilder()],
         super(key: key);
 
   /// [FocusNode] for the entire `SuperEditor`.
@@ -1206,6 +1207,8 @@ final defaultKeyboardActions = <DocumentKeyboardAction>[
   pasteWhenCmdVIsPressed,
   copyWhenCmdCIsPressed,
   cutWhenCmdXIsPressed,
+  undoWhenCmdZOrCtrlZIsPressed,
+  redoWhenCmdShiftZOrCtrlShiftZIsPressed,
   collapseSelectionWhenEscIsPressed,
   selectAllWhenCmdAIsPressed,
   moveLeftAndRightWithArrowKeys,
@@ -1215,12 +1218,19 @@ final defaultKeyboardActions = <DocumentKeyboardAction>[
   tabToIndentListItem,
   shiftTabToUnIndentListItem,
   backspaceToUnIndentListItem,
+  tabToIndentTask,
+  shiftTabToUnIndentTask,
+  backspaceToUnIndentTask,
+  tabToIndentParagraph,
+  shiftTabToUnIndentParagraph,
+  backspaceToUnIndentParagraph,
   backspaceToConvertTaskToParagraph,
   backspaceToClearParagraphBlockType,
   cmdBToToggleBold,
   cmdIToToggleItalics,
   shiftEnterToInsertNewlineInBlock,
   enterToInsertNewTask,
+  enterToUnIndentParagraph,
   enterToInsertBlockNewline,
   moveToLineStartOrEndWithCtrlAOrE,
   deleteToStartOfLineWithCmdBackspaceOnMac,
@@ -1247,12 +1257,22 @@ final defaultImeKeyboardActions = <DocumentKeyboardAction>[
   pasteWhenCmdVIsPressed,
   copyWhenCmdCIsPressed,
   cutWhenCmdXIsPressed,
+  undoWhenCmdZOrCtrlZIsPressed,
+  redoWhenCmdShiftZOrCtrlShiftZIsPressed,
   selectAllWhenCmdAIsPressed,
   cmdBToToggleBold,
   cmdIToToggleItalics,
   doNothingWithBackspaceOnWeb,
+  doNothingWithCtrlOrCmdAndZOnWeb,
+  tabToIndentTask,
+  shiftTabToUnIndentTask,
+  backspaceToUnIndentTask,
+  tabToIndentParagraph,
+  shiftTabToUnIndentParagraph,
+  backspaceToUnIndentParagraph,
   backspaceToConvertTaskToParagraph,
   backspaceToUnIndentListItem,
+  enterToUnIndentParagraph,
   backspaceToClearParagraphBlockType,
   deleteDownstreamCharacterWithCtrlDeleteOnMac,
   scrollOnCtrlOrCmdAndHomeKeyPress,
@@ -1494,6 +1514,14 @@ TextStyle defaultStyleBuilder(Set<Attribution> attributions) {
             ? TextDecoration.lineThrough
             : TextDecoration.combine([TextDecoration.lineThrough, newStyle.decoration!]),
       );
+    } else if (attribution == superscriptAttribution) {
+      newStyle = newStyle.copyWith(
+        fontFeatures: [const FontFeature.superscripts()],
+      );
+    } else if (attribution == subscriptAttribution) {
+      newStyle = newStyle.copyWith(
+        fontFeatures: [const FontFeature.subscripts()],
+      );
     } else if (attribution is ColorAttribution) {
       newStyle = newStyle.copyWith(
         color: attribution.color,
@@ -1505,6 +1533,10 @@ TextStyle defaultStyleBuilder(Set<Attribution> attributions) {
     } else if (attribution is FontSizeAttribution) {
       newStyle = newStyle.copyWith(
         fontSize: attribution.fontSize,
+      );
+    } else if (attribution is FontFamilyAttribution) {
+      newStyle = newStyle.copyWith(
+        fontFamily: attribution.fontFamily,
       );
     } else if (attribution is LinkAttribution) {
       newStyle = newStyle.copyWith(
