@@ -198,14 +198,39 @@ class TapSequenceGestureRecognizer extends GestureRecognizer {
     _trackers.remove(tracker.pointer);
     tracker.entry.resolve(GestureDisposition.rejected);
     _freezeTracker(tracker);
-    if (_firstTap != null || _secondTap != null) {
-      if (tracker == _firstTap || tracker == _secondTap) {
+
+    if (_firstTap == null && _firstTapDownDetails != null) {
+      // The user tapped down and then another recognizer won the arena. For example, in an app with both a
+      // TapSequenceGestureRecognizer and a HorizontalDragGestureRecognizer, when the user taps down and
+      // then drags horizontally, the onTapDown event is fired, and after that the HorizontalDragGestureRecognizer
+      // declares itself as the winner. Invoke onTapCancel to cancel the gesture.
+      _notifyListenersOfCancellation();
+      if (_trackers.isEmpty) {
         _reset();
-      } else {
-        _checkCancel();
-        if (_trackers.isEmpty) {
-          _reset();
-        }
+      }
+      return;
+    }
+
+    if (tracker == _secondTap) {
+      // A double tap was registered and we were defeated on the gesture arena after that. Reset
+      // to clean up the tap trackers.
+      _reset();
+      return;
+    }
+
+    if (tracker == _firstTap) {
+      // A tap up was registered and we were defeated on the gesture arena after that. Reset
+      // to clean up the tap trackers.
+      _reset();
+      return;
+    }
+
+    if (_firstTap != null || _secondTap != null) {
+      // We have a single or double tap registered, but the tracker isn't related to any of them.
+      // It's not clear what this situation means.
+      _notifyListenersOfCancellation();
+      if (_trackers.isEmpty) {
+        _reset();
       }
     }
   }
@@ -254,7 +279,7 @@ class TapSequenceGestureRecognizer extends GestureRecognizer {
     _stopTapTimer();
     if (_secondTap != null) {
       if (_trackers.isNotEmpty) {
-        _checkCancel();
+        _notifyListenersOfCancellation();
       }
       // Note, order is important below in order for the resolve -> reject logic
       // to work properly.
@@ -265,7 +290,7 @@ class TapSequenceGestureRecognizer extends GestureRecognizer {
     }
     if (_firstTap != null) {
       if (_trackers.isNotEmpty) {
-        _checkCancel();
+        _notifyListenersOfCancellation();
       }
       // Note, order is important below in order for the resolve -> reject logic
       // to work properly.
@@ -275,6 +300,8 @@ class TapSequenceGestureRecognizer extends GestureRecognizer {
       GestureBinding.instance.gestureArena.release(tracker.pointer);
     }
     _clearTrackers();
+
+    _firstTapDownDetails = null;
   }
 
   void _registerFirstTap(PointerEvent event, _TapTracker tracker) {
@@ -371,15 +398,26 @@ class TapSequenceGestureRecognizer extends GestureRecognizer {
     }
   }
 
-  void _checkCancel() {
-    if (_firstTap == null && onTapCancel != null) {
-      invokeCallback<void>('onTapCancel', onTapCancel!);
+  void _notifyListenersOfCancellation() {
+    if (_secondTap != null) {
+      if (onTripleTapCancel != null) {
+        invokeCallback<void>('onTripleTapCancel', onTripleTapCancel!);
+      }
+      return;
     }
-    if (_firstTap != null && _secondTap == null && onDoubleTapCancel != null) {
-      invokeCallback<void>('onDoubleTapCancel', onDoubleTapCancel!);
+
+    if (_firstTap != null) {
+      if (onDoubleTapCancel != null) {
+        invokeCallback<void>('onDoubleTapCancel', onDoubleTapCancel!);
+      }
+      return;
     }
-    if (_secondTap != null && onTripleTapCancel != null) {
-      invokeCallback<void>('onTripleTapCancel', onTripleTapCancel!);
+
+    if (_firstTapDownDetails != null) {
+      if (onTapCancel != null) {
+        invokeCallback<void>('onTapCancel', onTapCancel!);
+      }
+      return;
     }
   }
 
