@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:follow_the_leader/follow_the_leader.dart';
+import 'package:super_editor/src/default_editor/document_gestures_touch_ios.dart';
 import 'package:super_editor/src/infrastructure/flutter/flutter_scheduler.dart';
 import 'package:super_editor/src/infrastructure/multi_listenable_builder.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
@@ -429,16 +430,38 @@ class _IOSEditingControlsState extends State<IOSEditingControls>
     final upstreamTextPosition = selectionDirection == TextAffinity.downstream
         ? widget.editingController.textController.selection.base
         : widget.editingController.textController.selection.extent;
-    final upstreamHandleOffsetInText = _textPositionToTextOffset(upstreamTextPosition);
-    final upstreamLineHeight =
-        _textLayout.getCharacterBox(upstreamTextPosition)?.toRect().height ?? _textLayout.estimatedLineHeight;
 
     final downstreamTextPosition = selectionDirection == TextAffinity.downstream
         ? widget.editingController.textController.selection.extent
         : widget.editingController.textController.selection.base;
-    final downstreamHandleOffsetInText = _textPositionToTextOffset(downstreamTextPosition);
-    final downstreamLineHeight =
-        _textLayout.getCharacterBox(downstreamTextPosition)?.toRect().height ?? _textLayout.estimatedLineHeight;
+
+    late final Offset upstreamHandleOffsetInText;
+    late final double upstreamLineHeight;
+
+    late final Offset downstreamHandleOffsetInText;
+    late final double downstreamLineHeight;
+
+    final selectionBoxes = _textLayout.getBoxesForSelection(widget.editingController.textController.selection);
+    if (selectionBoxes.isEmpty) {
+      // It's not documented if getBoxesForSelection is guaranteed to return a non-empty list. Therefore,
+      // fallback to using character box to get the handle's offset and height.
+      upstreamHandleOffsetInText = _textPositionToTextOffset(upstreamTextPosition);
+      upstreamLineHeight =
+          _textLayout.getCharacterBox(upstreamTextPosition)?.toRect().height ?? _textLayout.estimatedLineHeight;
+
+      downstreamHandleOffsetInText = _textPositionToTextOffset(downstreamTextPosition);
+      downstreamLineHeight =
+          _textLayout.getCharacterBox(downstreamTextPosition)?.toRect().height ?? _textLayout.estimatedLineHeight;
+    } else {
+      final upstreamSelectionBox = selectionBoxes.first;
+      final downstreamSelectionBox = selectionBoxes.last;
+
+      upstreamHandleOffsetInText = Offset(upstreamSelectionBox.left, upstreamSelectionBox.top);
+      upstreamLineHeight = upstreamSelectionBox.bottom - upstreamSelectionBox.top;
+
+      downstreamHandleOffsetInText = Offset(downstreamSelectionBox.right, downstreamSelectionBox.top);
+      downstreamLineHeight = downstreamSelectionBox.bottom - downstreamSelectionBox.top;
+    }
 
     if (upstreamLineHeight == 0 || downstreamLineHeight == 0) {
       _log.finer('Not building expanded handles because the text layout reported a zero line-height');
@@ -484,12 +507,23 @@ class _IOSEditingControlsState extends State<IOSEditingControls>
     required Color debugColor,
     required void Function(DragStartDetails) onPanStart,
   }) {
+    final ballRadius = defaultIosHandleBallDiameter / 2;
+
     return CompositedTransformFollower(
       key: handleKey,
       link: widget.textContentLayerLink,
       offset: followerOffset,
       child: Transform.translate(
-        offset: const Offset(-12, -5),
+        offset: Offset(
+          -12,
+          -selectionHighlightBoxVerticalExpansion +
+              (isUpstreamHandle
+                  // For the upstream handle, the ball is displayed above the text, partially
+                  // overlapping the selected area. Move the ball up so it's positioned above the selected area,
+                  // and add half of the radius to make the ball overlap with the selected area.
+                  ? -defaultIosHandleBallDiameter + (ballRadius / 2)
+                  : 0),
+        ),
         child: TapRegion(
           groupId: widget.tapRegionGroupId,
           child: GestureDetector(
@@ -504,10 +538,12 @@ class _IOSEditingControlsState extends State<IOSEditingControls>
               child: showHandle
                   ? isUpstreamHandle
                       ? IOSSelectionHandle.upstream(
+                          ballRadius: ballRadius,
                           color: widget.handleColor,
                           caretHeight: lineHeight,
                         )
                       : IOSSelectionHandle.downstream(
+                          ballRadius: ballRadius,
                           color: widget.handleColor,
                           caretHeight: lineHeight,
                         )
