@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_test_robots/flutter_test_robots.dart';
@@ -9,7 +10,7 @@ import '../supereditor_test_tools.dart';
 
 void main() {
   group('SuperEditor link editing >', () {
-    group('recognizes a URL and converts it to a link', () {
+    group('recognizes a URL with https and www and converts it to a link', () {
       testWidgetsOnAllPlatforms('when typing', (tester) async {
         await tester //
             .createDocument()
@@ -1082,7 +1083,7 @@ void main() {
       });
     });
 
-    testWidgetsOnAllPlatforms('recognizes a second URL when typing and converts it to a link', (tester) async {
+    testWidgetsOnAllPlatforms('inserts https scheme if it is missing', (tester) async {
       await tester //
           .createDocument()
           .withSingleEmptyParagraph()
@@ -1092,35 +1093,30 @@ void main() {
       // Place the caret at the beginning of the empty document.
       await tester.placeCaretInParagraph("1", 0);
 
-      // Type text with two URLs
-      await tester.typeImeText("https://www.google.com and https://flutter.dev ");
+      // Type a URL. It shouldn't linkify until we add a space.
+      await tester.typeImeText("www.google.com");
 
-      // Ensure both URLs are linkified with the correct URLs.
-      final text = SuperEditorInspector.findTextInComponent("1");
+      // Type a space, to cause a linkify reaction.
+      await tester.typeImeText(" ");
 
-      expect(text.text, "https://www.google.com and https://flutter.dev ");
+      // Ensure it's linkified with a URL schema.
+      var text = SuperEditorInspector.findTextInComponent("1");
+      text = SuperEditorInspector.findTextInComponent("1");
+
+      expect(text.text, "www.google.com ");
       expect(
-        text.hasAttributionsThroughout(
-          attributions: {
-            LinkAttribution.fromUri(Uri.parse("https://www.google.com")),
-          },
-          range: const SpanRange(0, 21),
-        ),
-        isTrue,
-      );
-
-      expect(
-        text.hasAttributionsThroughout(
-          attributions: {
-            LinkAttribution.fromUri(Uri.parse("https://flutter.dev")),
-          },
-          range: const SpanRange(27, 45),
-        ),
-        isTrue,
+        text.getAttributionSpansByFilter((a) => a is LinkAttribution),
+        {
+          AttributionSpan(
+            attribution: LinkAttribution.fromUri(Uri.parse("https://www.google.com")),
+            start: 0,
+            end: 13,
+          ),
+        },
       );
     });
 
-    testWidgetsOnAllPlatforms('recognizes a URL without www and converts it to a link', (tester) async {
+    testWidgetsOnAllPlatforms('recognizes a URL without https and www and converts it to a link', (tester) async {
       await tester //
           .createDocument()
           .withSingleEmptyParagraph()
@@ -1153,17 +1149,18 @@ void main() {
 
       expect(text.text, "google.com ");
       expect(
-        text.hasAttributionsThroughout(
-          attributions: {
-            LinkAttribution.fromUri(Uri.parse("https://google.com")),
-          },
-          range: SpanRange(0, text.length - 2),
-        ),
-        isTrue,
+        text.getAttributionSpansByFilter((a) => a is LinkAttribution),
+        {
+          AttributionSpan(
+            attribution: LinkAttribution.fromUri(Uri.parse("https://google.com")),
+            start: 0,
+            end: 9,
+          ),
+        },
       );
     });
 
-    testWidgetsOnAllPlatforms('inserts https scheme if it is missing', (tester) async {
+    testWidgetsOnAllPlatforms('recognizes a second URL when typing and converts it to a link', (tester) async {
       await tester //
           .createDocument()
           .withSingleEmptyParagraph()
@@ -1173,25 +1170,134 @@ void main() {
       // Place the caret at the beginning of the empty document.
       await tester.placeCaretInParagraph("1", 0);
 
-      // Type a URL. It shouldn't linkify until we add a space.
-      await tester.typeImeText("www.google.com");
+      // Type text with two URLs.
+      await tester.typeImeText("https://www.google.com and https://flutter.dev ");
 
-      // Type a space, to cause a linkify reaction.
-      await tester.typeImeText(" ");
+      // Ensure both URLs are linkified with the correct URLs.
+      final text = SuperEditorInspector.findTextInComponent("1");
+
+      expect(text.text, "https://www.google.com and https://flutter.dev ");
+      expect(
+        text.getAttributionSpansByFilter((a) => a is LinkAttribution),
+        {
+          AttributionSpan(
+            attribution: LinkAttribution.fromUri(Uri.parse("https://www.google.com")),
+            start: 0,
+            end: 21,
+          ),
+          AttributionSpan(
+            attribution: LinkAttribution.fromUri(Uri.parse("https://flutter.dev")),
+            start: 27,
+            end: 45,
+          ),
+        },
+      );
+    });
+
+    testWidgetsOnDesktop('recognizes a pasted URL with www and converts it to a link', (tester) async {
+      await tester //
+          .createDocument()
+          .withSingleEmptyParagraph()
+          .withInputSource(TextInputSource.ime)
+          .pump();
+
+      // Place the caret at the beginning of the empty document.
+      await tester.placeCaretInParagraph("1", 0);
+
+      // Paste text with a URL.
+      tester.simulateClipboard();
+      await tester.setSimulatedClipboardContent("Hello https://www.google.com world");
+      // TODO: create and use something like tester.pressPasteAdaptive()
+      if (debugDefaultTargetPlatformOverride == TargetPlatform.macOS) {
+        await tester.pressCmdV();
+      } else {
+        await tester.pressCtlV();
+      }
+
+      // Ensure the URL is linkified.
+      final text = SuperEditorInspector.findTextInComponent("1");
+      expect(text.text, "Hello https://www.google.com world");
+      expect(
+        text.getAttributionSpansByFilter((a) => a is LinkAttribution),
+        {
+          AttributionSpan(
+            attribution: LinkAttribution.fromUri(Uri.parse("https://www.google.com")),
+            start: 6,
+            end: 27,
+          ),
+        },
+      );
+    });
+
+    testWidgetsOnDesktop('recognizes a pasted URL and inserts https scheme if it is missing', (tester) async {
+      await tester //
+          .createDocument()
+          .withSingleEmptyParagraph()
+          .withInputSource(TextInputSource.ime)
+          .pump();
+
+      // Place the caret at the beginning of the empty document.
+      await tester.placeCaretInParagraph("1", 0);
+
+      // Paste text with a URL.
+      tester.simulateClipboard();
+      await tester.setSimulatedClipboardContent("Hello www.google.com world");
+      // TODO: create and use something like tester.pressPasteAdaptive()
+      if (debugDefaultTargetPlatformOverride == TargetPlatform.macOS) {
+        await tester.pressCmdV();
+      } else {
+        await tester.pressCtlV();
+      }
 
       // Ensure it's linkified with a URL schema.
       var text = SuperEditorInspector.findTextInComponent("1");
       text = SuperEditorInspector.findTextInComponent("1");
 
-      expect(text.text, "www.google.com ");
+      expect(text.text, "Hello www.google.com world");
       expect(
-        text.hasAttributionsThroughout(
-          attributions: {
-            LinkAttribution.fromUri(Uri.parse("https://www.google.com")),
-          },
-          range: SpanRange(0, text.length - 2),
-        ),
-        isTrue,
+        text.getAttributionSpansByFilter((a) => a is LinkAttribution),
+        {
+          AttributionSpan(
+            attribution: LinkAttribution.fromUri(Uri.parse("https://www.google.com")),
+            start: 6,
+            end: 19,
+          ),
+        },
+      );
+    });
+
+    testWidgetsOnDesktop('recognizes a pasted URL without https or www and converts it to a link', (tester) async {
+      await tester //
+          .createDocument()
+          .withSingleEmptyParagraph()
+          .withInputSource(TextInputSource.ime)
+          .pump();
+
+      // Place the caret at the beginning of the empty document.
+      await tester.placeCaretInParagraph("1", 0);
+
+      // Paste text with a URL.
+      tester.simulateClipboard();
+      await tester.setSimulatedClipboardContent("Hello google.com world");
+      // TODO: create and use something like tester.pressPasteAdaptive()
+      if (debugDefaultTargetPlatformOverride == TargetPlatform.macOS) {
+        await tester.pressCmdV();
+      } else {
+        await tester.pressCtlV();
+      }
+
+      // Ensure the URL is linkified.
+      final text = SuperEditorInspector.findTextInComponent("1");
+      expect(text.text, "Hello google.com world");
+      expect(
+        text.getAttributionSpansByFilter((a) => a is LinkAttribution),
+        {
+          AttributionSpan(
+            attribution: LinkAttribution.fromUri(Uri.parse("https://google.com")),
+            start: 6,
+            end: 15,
+          ),
+        },
       );
     });
 
@@ -1211,7 +1317,7 @@ void main() {
         // Type some text by simulating hardware keyboard key presses.
         await tester.typeKeyboardText('Go to ');
 
-        // Ensure that the link is unchanged
+        // Ensure that the link is unchanged.
         expect(
           SuperEditorInspector.findDocument(),
           equalsMarkdown("Go to [www.google.com](www.google.com)"),
@@ -1234,7 +1340,7 @@ void main() {
         // Type some text by simulating hardware keyboard key presses.
         await tester.typeKeyboardText('Go to ');
 
-        // Ensure that the link is unchanged
+        // Ensure that the link is unchanged.
         expect(
           SuperEditorInspector.findDocument(),
           equalsMarkdown("Go to [www.google.com](www.google.com)"),
@@ -1257,7 +1363,7 @@ void main() {
         // Type some text by simulating hardware keyboard key presses.
         await tester.typeKeyboardText('Go to ');
 
-        // Ensure that the link is unchanged
+        // Ensure that the link is unchanged.
         expect(
           SuperEditorInspector.findDocument(),
           equalsMarkdown("Go to [www.google.com](www.google.com)"),
@@ -1281,7 +1387,7 @@ void main() {
         // Type some text by simulating hardware keyboard key presses.
         await tester.typeKeyboardText(' to learn anything');
 
-        // Ensure that the link is unchanged
+        // Ensure that the link is unchanged.
         expect(
           SuperEditorInspector.findDocument(),
           equalsMarkdown("[www.google.com](www.google.com) to learn anything"),
@@ -1304,7 +1410,7 @@ void main() {
         // Type some text by simulating hardware keyboard key presses.
         await tester.typeKeyboardText(' to learn anything');
 
-        // Ensure that the link is unchanged
+        // Ensure that the link is unchanged.
         expect(
           SuperEditorInspector.findDocument(),
           equalsMarkdown("[www.google.com](www.google.com) to learn anything"),
@@ -1327,7 +1433,7 @@ void main() {
         // Type some text by simulating hardware keyboard key presses.
         await tester.typeKeyboardText(' to learn anything');
 
-        // Ensure that the link is unchanged
+        // Ensure that the link is unchanged.
         expect(
           SuperEditorInspector.findDocument(),
           equalsMarkdown("[www.google.com](www.google.com) to learn anything"),
@@ -1377,7 +1483,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.goog|le.com"
+        // Place the caret at "www.goog|le.com".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 8);
 
         // Add characters.
@@ -1407,7 +1513,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.goog|le.com"
+        // Place the caret at "www.goog|le.com".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 8);
 
         // Add characters.
@@ -1430,7 +1536,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "|www.google.com"
+        // Place the caret at "|www.google.com".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 0);
 
         // Delete downstream characters.
@@ -1465,7 +1571,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "|www.google.com"
+        // Place the caret at "|www.google.com".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 0);
 
         // Delete downstream characters.
@@ -1523,7 +1629,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "|www.google.com"
+        // Place the caret at "|www.google.com".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 0);
 
         // Delete downstream characters.
@@ -1549,7 +1655,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.google.com|"
+        // Place the caret at "www.google.com|".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 10);
 
         // Delete upstream characters.
@@ -1583,7 +1689,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.google|.com"
+        // Place the caret at "www.google|.com".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 10);
 
         // Remove characters.
@@ -1621,7 +1727,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.google|.com"
+        // Place the caret at "www.google|.com".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 10);
 
         // Remove a single character.
@@ -1644,7 +1750,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.google.com|"
+        // Place the caret at "www.google.com|".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 14);
 
         // Delete upstream characters.
@@ -1679,7 +1785,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.google.com|"
+        // Place the caret at "www.google.com|".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 14);
 
         // Delete upstream characters.
@@ -1710,7 +1816,7 @@ void main() {
 
         final doc = SuperEditorInspector.findDocument()!;
 
-        // Place the caret at "www.google.com|"
+        // Place the caret at "www.google.com|".
         await tester.placeCaretInParagraph(doc.nodes.first.id, 14);
 
         // Delete an upstream characters.
@@ -1815,7 +1921,7 @@ void main() {
 
       final doc = SuperEditorInspector.findDocument()!;
 
-      // Place the caret at "www.google.com|"
+      // Place the caret at "www.google.com|".
       await tester.placeCaretInParagraph(doc.nodes.first.id, 14);
 
       // Delete a character at the end of the link.
@@ -1926,6 +2032,56 @@ void main() {
           range: SpanRange(0, newListItemText.length - 1),
         ),
         isEmpty,
+      );
+    });
+
+    testWidgetsOnAllPlatforms('plays nice with Markdown link when Markdown parsing is disabled', (tester) async {
+      // Based on bug #2074 - https://github.com/superlistapp/super_editor/issues/2074
+      await tester //
+          .createDocument()
+          .withSingleEmptyParagraph()
+          .withInputSource(TextInputSource.ime)
+          .pump();
+
+      await tester.placeCaretInParagraph("1", 0);
+
+      await tester.typeImeText("[google](www.google.com) ");
+
+      // Ensure that the Markdown was ignored and nothing was linkified.
+      final text = SuperEditorInspector.findTextInComponent("1");
+      expect(text.text, "[google](www.google.com) ");
+      expect(text.getAttributionSpansByFilter((a) => true), isEmpty);
+    });
+
+    testWidgetsOnMac('plays nice with Markdown link when pasting a Markdown link', (tester) async {
+      // Based on bug #2074 - https://github.com/superlistapp/super_editor/issues/2074
+      await tester //
+          .createDocument()
+          .withSingleEmptyParagraph()
+          .withInputSource(TextInputSource.ime)
+          .pump();
+
+      await tester.placeCaretInParagraph("1", 0);
+
+      // Simulate copying a Markdown link to the clipboard.
+      tester.simulateClipboard();
+      await tester.setSimulatedClipboardContent("Hello [google](www.google.com) ");
+
+      // Simulate pasting the Markdown link into the document.
+      await tester.pressCmdV();
+
+      // Ensure that the Markdown was ignored and nothing was linkified.
+      final text = SuperEditorInspector.findTextInComponent("1");
+      expect(text.text, "Hello [google](www.google.com) ");
+      expect(text.getAttributionSpansByFilter((a) => true), isEmpty);
+      expect(
+        SuperEditorInspector.findDocumentSelection(),
+        const DocumentSelection.collapsed(
+          position: DocumentPosition(
+            nodeId: "1",
+            nodePosition: TextNodePosition(offset: 31),
+          ),
+        ),
       );
     });
 
