@@ -10,6 +10,25 @@ import '../../test_documents.dart';
 void main() {
   group("SuperEditor stable tags >", () {
     group("composing >", () {
+      testWidgetsOnAllPlatforms("starts with a trigger", (tester) async {
+        await _pumpTestEditor(
+          tester,
+          singleParagraphEmptyDoc(),
+        );
+        await tester.placeCaretInParagraph("1", 0);
+
+        // Compose a stable tag.
+        await tester.typeImeText("@");
+
+        // Ensure that the tag has a composing attribution.
+        final text = SuperEditorInspector.findTextInComponent("1");
+        expect(text.text, "@");
+        expect(
+          text.getAttributedRange({stableTagComposingAttribution}, 0),
+          const SpanRange(0, 0),
+        );
+      });
+
       testWidgetsOnAllPlatforms("can start at the beginning of a paragraph", (tester) async {
         await _pumpTestEditor(
           tester,
@@ -119,9 +138,6 @@ void main() {
               ),
             ],
           ),
-          plugin: StableTagPlugin(
-            tagRule: const TagRule(trigger: '@'),
-          ),
         );
 
         // Place the caret at "before |"
@@ -144,8 +160,14 @@ void main() {
         text = SuperEditorInspector.findTextInComponent("1");
         expect(text.text, "before @john after");
         expect(
-          text.getAttributedRange({stableTagComposingAttribution}, 7),
-          const SpanRange(7, 17),
+          text.getAttributionSpansByFilter((a) => a == stableTagComposingAttribution),
+          {
+            const AttributionSpan(
+              attribution: stableTagComposingAttribution,
+              start: 7,
+              end: 17,
+            ),
+          },
         );
       });
 
@@ -1107,25 +1129,19 @@ void main() {
     });
 
     group("emoji >", () {
-      testWidgetsOnAllPlatforms("typed as first character of a paragraph", (tester) async {
+      testWidgetsOnAllPlatforms("can be typed as first character of a paragraph without crashing the editor",
+          (tester) async {
         // Ensure we can type an emoji as first character without crashing
         // https://github.com/superlistapp/super_editor/issues/1863
         await _pumpTestEditor(
           tester,
-          MutableDocument(
-            nodes: [
-              ParagraphNode(
-                id: "1",
-                text: AttributedText(""),
-              ),
-            ],
-          ),
+          singleParagraphEmptyDoc(),
         );
 
         // Place the caret at the beginning of the paragraph.
         await tester.placeCaretInParagraph("1", 0);
 
-        // Type an emoji as first charactet 💙
+        // Type an emoji as first character 💙
         await tester.typeImeText("💙");
 
         expect(
@@ -1142,8 +1158,8 @@ void main() {
         expect(text.text, "💙");
       });
 
-      testWidgetsOnAllPlatforms("after trigger", (tester) async {
-        final plugin = StableTagPlugin();
+      testWidgetsOnAllPlatforms("caret can move around emoji without breaking editor", (tester) async {
+        // We are doing this to ensure the plugin doesn't make the editor crash when moving caret around emoji.
         await _pumpTestEditor(
           tester,
           MutableDocument(
@@ -1154,13 +1170,25 @@ void main() {
               ),
             ],
           ),
-          plugin: plugin,
         );
 
-        // Place the caret after the emoji: 💙|
-        await tester.placeCaretInParagraph("1", 2);
+        // Place the caret before the emoji: |💙
+        await tester.placeCaretInParagraph("1", 0);
 
-        // Move the caret before the emoji: |💙
+        // Place the caret after the emoji: 💙|
+        await tester.pressRightArrow();
+
+        expect(
+          SuperEditorInspector.findDocumentSelection(),
+          const DocumentSelection.collapsed(
+            position: DocumentPosition(
+              nodeId: "1",
+              nodePosition: TextNodePosition(offset: 2),
+            ),
+          ),
+        );
+
+        // Move the caret back to initial position, before the emoji: |💙.
         await tester.pressLeftArrow();
 
         expect(
@@ -1173,49 +1201,58 @@ void main() {
           ),
         );
 
-        // Type @ and ensure the string is valid: @💙
-        await tester.typeImeText("@");
+        // Ensure the paragraph string is well formed: 💙
+        final text = SuperEditorInspector.findTextInComponent("1");
+        expect(text.text, "💙");
+      });
+
+      testWidgetsOnAllPlatforms("can be captured with trigger", (tester) async {
+        await _pumpTestEditor(
+          tester,
+          singleParagraphEmptyDoc(),
+        );
+
+        // Place the caret at the beginning of the paragraph.
+        await tester.placeCaretInParagraph("1", 0);
+
+        // Type @ to trigger a composing tag, followed by an emoji 💙
+        await tester.typeImeText("@💙");
+
+        // Ensure the emoji is in the tag, and nothing went wrong with string formation.
         final text = SuperEditorInspector.findTextInComponent("1");
         expect(text.text, "@💙");
 
-        // Ensure the composing tag is on the emoji
-        final composingStableTag = plugin.tagIndex.composingStableTag.value!;
+        // Ensure the composing tag includes the emoji.
         expect(
-          composingStableTag,
-          const ComposingStableTag(
-            DocumentRange(
-              start: DocumentPosition(
-                nodeId: '1',
-                nodePosition: TextNodePosition(offset: 1),
-              ),
-              end: DocumentPosition(
-                nodeId: '1',
-                nodePosition: TextNodePosition(offset: 3),
-              ),
+          SuperEditorInspector.findTextInComponent("1")
+              .getAttributionSpansByFilter((a) => a == stableTagComposingAttribution),
+          {
+            const AttributionSpan(
+              attribution: stableTagComposingAttribution,
+              start: 0,
+              end: 2,
             ),
-            '💙',
-          ),
+          },
         );
 
-        // Move the caret after the emoji: @💙|
-        await tester.pressRightArrow();
-
-        // Type a white space to commit the tag: @💙 |
+        // Commit the tag.
         await tester.typeImeText(" ");
 
-        final commitedTags = plugin.tagIndex.getCommittedTagsInTextNode('1');
-
-        expect(commitedTags.length, 1);
-
-        final commitTag = commitedTags.first;
-
         // Ensure the committed tag is the emoji and the composing tag is removed
-        expect(commitTag, const IndexedTag(Tag('@', '💙'), '1', 0));
-        expect(plugin.tagIndex.composingStableTag.value, isNull);
+        expect(
+          SuperEditorInspector.findTextInComponent("1")
+              .getAttributionSpansByFilter((a) => a is CommittedStableTagAttribution),
+          {
+            const AttributionSpan(
+              attribution: CommittedStableTagAttribution("💙"),
+              start: 0,
+              end: 2,
+            ),
+          },
+        );
       });
 
-      testWidgetsOnAllPlatforms("before trigger", (tester) async {
-        final plugin = StableTagPlugin();
+      testWidgetsOnAllPlatforms("can be used before a trigger", (tester) async {
         await _pumpTestEditor(
           tester,
           MutableDocument(
@@ -1226,85 +1263,56 @@ void main() {
               ),
             ],
           ),
-          plugin: plugin,
         );
 
         // Place the caret after the emoji: 💙|
         await tester.placeCaretInParagraph("1", 2);
 
-        // Type @ and ensure the string is valid: 💙@
+        // Type @ to trigger a composing tag: 💙@
         await tester.typeImeText("@");
 
+        // Ensure nothing went wrong with the string construction.
         final text = SuperEditorInspector.findTextInComponent("1");
         expect(text.text, "💙@");
 
-        // Ensure the composing tag is not null and empty
+        // Ensure the tag was committed with the emoji.
         expect(
-          plugin.tagIndex.composingStableTag.value,
-          const ComposingStableTag(
-            DocumentRange(
-              start: DocumentPosition(
-                nodeId: '1',
-                nodePosition: TextNodePosition(offset: 3),
-              ),
-              end: DocumentPosition(
-                nodeId: '1',
-                nodePosition: TextNodePosition(offset: 3),
-              ),
+          SuperEditorInspector.findTextInComponent("1")
+              .getAttributionSpansByFilter((a) => a == stableTagComposingAttribution),
+          {
+            const AttributionSpan(
+              attribution: stableTagComposingAttribution,
+              start: 2,
+              end: 2,
             ),
-            '',
-          ),
+          },
         );
       });
 
-      testWidgetsOnAllPlatforms("in the middle of a tag", (tester) async {
-        final plugin = StableTagPlugin();
+      testWidgetsOnAllPlatforms("can be used in the middle of a tag", (tester) async {
         await _pumpTestEditor(
           tester,
-          MutableDocument(
-            nodes: [
-              ParagraphNode(
-                id: "1",
-                text: AttributedText(""),
-              ),
-            ],
-          ),
-          plugin: plugin,
+          singleParagraphEmptyDoc(),
         );
 
         // Place the caret at the beginning of the paragraph.
         await tester.placeCaretInParagraph("1", 0);
 
-        // Start compsing a tag with an emoji in the middle
-        await tester.typeImeText("@Flutter💙SuperEditor");
+        // Start composing a tag with an emoji in the middle
+        await tester.typeImeText("@Flutter💙SuperEditor ");
 
-        // Ensure the composing tag is valid
-        final composingStableTag = plugin.tagIndex.composingStableTag.value!;
+        // Ensure the tag was committed with the emoji.
         expect(
-          composingStableTag,
-          const ComposingStableTag(
-            DocumentRange(
-              start: DocumentPosition(
-                nodeId: '1',
-                nodePosition: TextNodePosition(offset: 1),
-              ),
-              end: DocumentPosition(
-                nodeId: '1',
-                nodePosition: TextNodePosition(offset: 21),
-              ),
+          SuperEditorInspector.findTextInComponent("1")
+              .getAttributionSpansByFilter((a) => a is CommittedStableTagAttribution),
+          {
+            const AttributionSpan(
+              attribution: CommittedStableTagAttribution("Flutter💙SuperEditor"),
+              start: 0,
+              end: 20,
             ),
-            'Flutter💙SuperEditor',
-          ),
+          },
         );
-
-        // Commit the tag by typing a white space
-        await tester.typeImeText(" ");
-
-        // Ensure the tag is commited and the composing tag is reset
-        final commitedTags = plugin.tagIndex.getCommittedTagsInTextNode('1');
-        expect(commitedTags.length, 1);
-        expect(commitedTags.first, const IndexedTag(Tag('@', 'Flutter💙SuperEditor'), '1', 0));
-        expect(plugin.tagIndex.composingStableTag.value, isNull);
       });
     });
   });
