@@ -9,9 +9,25 @@ import 'package:super_editor/super_text_field.dart';
 import '../super_textfield_inspector.dart';
 
 void main() {
-  group("SuperTextField > selection >", () {
+  group("SuperTextField mobile selection > iOS", () {
+    group("on tap >", () {
+      testWidgetsOnIos("when beyond first character > places caret at end of word", (tester) async {
+        // TODO: Add this test - for an example, see the Super Editor version: super_editor_ios_selection_test.dart
+        //       This test isn't implemented because when I got to it we didn't have any WidgetTester
+        //       extensions to tap to place the caret. Create those extensions and then implement this.
+        //       Issue: https://github.com/superlistapp/super_editor/issues/2098
+      }, skip: true);
+
+      testWidgetsOnIos("when near first character > places caret at start of word", (tester) async {
+        // TODO: Add this test - for an example, see the Super Editor version: super_editor_ios_selection_test.dart
+        //       This test isn't implemented because when I got to it we didn't have any WidgetTester
+        //       extensions to tap to place the caret. Create those extensions and then implement this.
+        //       Issue: https://github.com/superlistapp/super_editor/issues/2098
+      }, skip: true);
+    });
+
     testWidgetsOnIos("tapping on caret toggles the toolbar", (tester) async {
-      await _pumpTestApp(tester);
+      await _pumpScaffold(tester);
 
       // Ensure there's no selection to begin with, and no toolbar is displayed.
       expect(SuperTextFieldInspector.findSelection(), const TextSelection.collapsed(offset: -1));
@@ -40,20 +56,20 @@ void main() {
   });
 }
 
-Future<void> _pumpTestApp(
+Future<void> _pumpScaffold(
   WidgetTester tester, {
   AttributedTextEditingController? controller,
   EdgeInsets? padding,
   TextAlign? textAlign,
 }) async {
   final textFieldFocusNode = FocusNode();
-  const tapRegionGroupdId = "test_super_text_field";
+  const tapRegionGroupId = "test_super_text_field";
 
   await tester.pumpWidget(
     MaterialApp(
       home: Scaffold(
         body: TapRegion(
-          groupId: tapRegionGroupdId,
+          groupId: tapRegionGroupId,
           onTapOutside: (_) {
             // Unfocus on tap outside so that we're sure that all gesture tests
             // pass when using TapRegion's for focus, because apps should be able
@@ -71,7 +87,7 @@ Future<void> _pumpTestApp(
                   ),
                   child: SuperTextField(
                     focusNode: textFieldFocusNode,
-                    tapRegionGroupId: tapRegionGroupdId,
+                    tapRegionGroupId: tapRegionGroupId,
                     padding: padding,
                     textAlign: textAlign ?? TextAlign.left,
                     textController: controller ??
