@@ -144,7 +144,7 @@ void main() {
 
         await expectLater(
           find.byType(MaterialApp).first,
-          matchesGoldenFileWithPixelAllowance("goldens/text-scaling-header.png", 90),
+          matchesGoldenFileWithPixelAllowance("goldens/text-scaling-header.png", 125),
         );
       });
 
@@ -163,7 +163,7 @@ void main() {
 
         await expectLater(
           find.byType(MaterialApp).first,
-          matchesGoldenFileWithPixelAllowance("goldens/text-scaling-blockquote.png", 31),
+          matchesGoldenFileWithPixelAllowance("goldens/text-scaling-blockquote.png", 40),
         );
       });
     });
